run&jump WIP

+++ src/main/java/com/reve/thirst/thirst/Thirst.java
@@ -4,15 +4,15 @@ import java.util.HashMap;
 import java.util.UUID;
 
 public class Thirst {
-    public static HashMap<UUID, Integer> thirst = new HashMap<>();
-    public static final int DEFAULT = 20;
+    public static HashMap<UUID, Float> thirst = new HashMap<>();
+    public static final float DEFAULT = 20L;
     public Thirst(UUID id){
         thirst.put(id, DEFAULT);
     }
-    public static int getThirst(UUID id){
+    public static Float getThirst(UUID id){
         return thirst.get(id);
     }
-    public static void setThirst(UUID id, int value){
+    public static void setThirst(UUID id, float value){
         if (thirst.containsKey(id)) thirst.replace(id, value);
         else thirst.put(id, DEFAULT);
     }
